Use MUnit in StreamFlowWindowSuite
